+++ package.json
@@ -15,10 +15,6 @@
   "dependencies": {
     "fwk": "0.1.x"
   },
-  "devDependencies": {
-    "chai" : "0.2.4",
-    "mocha" : "0.10.x"
-  },                  
   "main" : "./lib/ddb",
   "engines" : { "node" : ">=0.6.0" }
 }
